typo in r27953 spotted by kugel: s/long/ssize_t/
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@27955 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/common/format.c
@@ -173,7 +173,7 @@ int format(
             ch = *fmt++;
             switch(ch) {
                 case 'd':
-                    szval = szsign = va_arg (ap, long);
+                    szval = szsign = va_arg (ap, ssize_t);
                     if (szval < 0)
                         szval = -szval;
                     do
